requirements.txt
kafka-python
scikit-learn
pandas
joblib
